Reorganization and renaming from David's comments
